Refactored bq_helper

+++ src/utils/bq_helper.py
@@ -21,10 +21,11 @@ def delete_table(dataset_id: str, table_ids: list):
     try:
         storage_client = bigquery.Client()
         for id in table_ids:
-            table_ref = storage_client.dataset(dataset_id).table(id)
-            # API request
-            logger.info('Deleting table {}:{}'.format(dataset_id, id))
-            storage_client.delete_table(table_ref)
+            if id != '_sessionId_update':
+                table_ref = storage_client.dataset(dataset_id).table(id)
+                # API request
+                logger.info('Deleting table {}:{}'.format(dataset_id, id))
+                storage_client.delete_table(table_ref)
     except Exception:
         raise ValueError('Desired table does not exist')
